Migrate test_sed_kern to test_sed_kernels.py
