+++ diffsky/experimental/kernels/tests/test_mc_phot_kernels.py
@@ -1,7 +1,6 @@
 """"""
 
 import numpy as np
-from diffstar import DEFAULT_DIFFSTAR_PARAMS
 from dsps.cosmology import DEFAULT_COSMOLOGY
 from dsps.data_loaders import load_emline_info as lemi
 from dsps.photometry import photometry_kernels as phk
@@ -18,63 +17,6 @@ from .. import mc_phot_kernels as mcpk
 
 _A = [None, 0, None, None, 0, *[None] * 4]
 calc_obs_mags_galpop = vmap(phk.calc_obs_mag, in_axes=_A)
-
-
-def test_sed_kern(num_halos=250):
-    ran_key = jran.key(0)
-    lc_data, tcurves = tmclh._get_weighted_lc_data_for_unit_testing(num_halos=num_halos)
-
-    fb = 0.156
-    ran_key, phot_key = jran.split(ran_key, 2)
-
-    phot_kern_results, phot_randoms = mcpk._mc_phot_kern(
-        phot_key,
-        lc_data.z_obs,
-        lc_data.t_obs,
-        lc_data.mah_params,
-        lc_data.ssp_data,
-        lc_data.precomputed_ssp_mag_table,
-        lc_data.z_phot_table,
-        lc_data.wave_eff_table,
-        *dpw.DEFAULT_PARAM_COLLECTION,
-        DEFAULT_COSMOLOGY,
-        fb,
-    )
-    sfh_params = DEFAULT_DIFFSTAR_PARAMS._make(
-        [getattr(phot_kern_results, key) for key in DEFAULT_DIFFSTAR_PARAMS._fields]
-    )
-    sed_kern_results = mcpk._sed_kern(
-        phot_randoms,
-        sfh_params,
-        lc_data.z_obs,
-        lc_data.t_obs,
-        lc_data.mah_params,
-        lc_data.ssp_data,
-        *dpw.DEFAULT_PARAM_COLLECTION[1:],
-        DEFAULT_COSMOLOGY,
-        fb,
-    )
-    rest_sed_recomputed = sed_kern_results[0]
-
-    # Enforce agreement between precomputed vs exact magnitudes
-    n_bands = phot_kern_results.obs_mags.shape[1]
-    for iband in range(n_bands):
-        trans_iband = np.interp(
-            lc_data.ssp_data.ssp_wave,
-            tcurves[iband].wave,
-            tcurves[iband].transmission,
-        )
-        args = (
-            lc_data.ssp_data.ssp_wave,
-            rest_sed_recomputed,
-            lc_data.ssp_data.ssp_wave,
-            trans_iband,
-            lc_data.z_obs,
-            *DEFAULT_COSMOLOGY,
-        )
-
-        mags = calc_obs_mags_galpop(*args)
-        assert np.allclose(mags, phot_kern_results.obs_mags[:, iband], rtol=0.01)
 
 
 def test_mc_dbk_kern(num_halos=50):
